branch:      naev [svn] lasers fade out minor tweaks fixed all memleaks just need some collision detection now

+++ src/rng.h
@@ -7,7 +7,7 @@
 #include <stdlib.h>
 
 
-#define RNG(L,H)	(rand()%(H-L+1)+L)
+#define RNG(L,H)	(rand()%(int)(H-L+1)+L)
 
 
 void rng_init (void);
